# backend/src/main/resources/api-spec.yaml
openapi: 3.0.0
info:
  title: Blog System API
  description: An API for a blog system that allows users to create posts and comment on them.
  version: "1.0.0"
servers:
  - url: 'https://example.com/api'
paths:
  /users:
    get:
      summary: Lists all users
      operationId: listUsers
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
  /posts:
    get:
      summary: Lists all posts
      operationId: listPosts
      responses:
        '200':
          description: A list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Create a new post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '201':
          description: Post created
  /posts/{postId}/comments:
    get:
      summary: Lists all comments for a post
      operationId: listCommentsByPost
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of comments for the post
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
    post:
      summary: Create a new comment on a post
      operationId: createComment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        '201':
          description: Comment created
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        username:
          type: string
        email:
          type: string
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
        content:
          type: string
        userId:
          type: integer
          format: int64
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        content:
          type: string
        postId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64